Move and rename index.rs to index/usearch.rs
