+++ src/index/opensearch.rs
@@ -1,0 +1,4 @@
+/*
+ * Copyright 2025-present ScyllaDB
+ * SPDX-License-Identifier: Apache-2.0
+ */
